+++ client/src/app/auth/tabs/dashboard/accounts/accounts-configuration-card.tsx
