+++ client/src/app/auth/tabs/dashboard/accounts/delete-account.tsx
