+++ client/src/app/auth/tabs/dashboard/accounts/deleted-account-card.tsx
