+++ client/src/app/auth/tabs/dashboard/sync-account-button.tsx
