+++ client/src/app/auth/tabs/transactions/deleted-transaction-card.tsx
