+++ client/src/components/email-verified.tsx
